@@ -1,14 +1,18 @@
 from collections import defaultdict
-from typing import Any, Sequence, TypeVar
+from collections.abc import Sequence
+from typing import Any
 
 from shrinkray.passes.definitions import ReductionPass
 from shrinkray.passes.patching import CutPatch, Cuts, apply_patches
 from shrinkray.problem import ReductionProblem
 
-Seq = TypeVar("Seq", bound=Sequence[Any])
 
+async def delete_elements[Seq: Sequence[Any]](problem: ReductionProblem[Seq]) -> None:
+    """Try to delete individual elements from the sequence.
 
-async def delete_elements(problem: ReductionProblem[Seq]) -> None:
+    Creates a patch for each element and uses the patch applier to find
+    which elements can be removed while maintaining interestingness.
+    """
     await apply_patches(
         problem, Cuts(), [[(i, i + 1)] for i in range(len(problem.current_test_case))]
     )
@@ -24,7 +28,9 @@ def merged_intervals(intervals: list[tuple[int, int]]) -> list[tuple[int, int]]:
     return list(map(tuple, normalized))  # type: ignore
 
 
-def with_deletions(target: Seq, deletions: list[tuple[int, int]]) -> Seq:
+def with_deletions[Seq: Sequence[Any]](
+    target: Seq, deletions: list[tuple[int, int]]
+) -> Seq:
     result: list[Any] = []
     prev = 0
     total_deleted = 0
@@ -37,7 +43,15 @@ def with_deletions(target: Seq, deletions: list[tuple[int, int]]) -> Seq:
     return type(target)(result)  # type: ignore
 
 
-def block_deletion(min_block: int, max_block: int) -> ReductionPass[Seq]:
+def block_deletion[Seq: Sequence[Any]](
+    min_block: int, max_block: int
+) -> ReductionPass[Seq]:
+    """Create a pass that deletes contiguous blocks of elements.
+
+    Tries to remove blocks of size min_block to max_block, starting at
+    various offsets. Useful for removing larger chunks efficiently.
+    """
+
     async def apply(problem: ReductionProblem[Seq]) -> None:
         n = len(problem.current_test_case)
         if n <= min_block:
@@ -55,7 +69,13 @@ def block_deletion(min_block: int, max_block: int) -> ReductionPass[Seq]:
     return apply
 
 
-async def delete_duplicates(problem: ReductionProblem[Seq]) -> None:
+async def delete_duplicates[Seq: Sequence[Any]](problem: ReductionProblem[Seq]) -> None:
+    """Try to delete duplicate elements from the sequence.
+
+    Groups elements by value and tries to remove all occurrences of each
+    duplicated element together. Effective when the test case contains
+    repeated patterns that can be eliminated.
+    """
     index: dict[int, list[int]] = defaultdict(list)
 
     for i, c in enumerate(problem.current_test_case):
